Fix SplitZip when no secondary dexs are created.
Summary:
When D773417 moved us to Guava, it changed ZipWithMaxDeflate to close instead of closeQuietly.  Native M Site and Appcenter both request to be splitzipped, even though they don't really need it for all builds.
This resulted in SplitZip producing no secondary dexes, which we then tried to zip with ZipWithMaxDeflate, which crashed when we tried to produce an emtpy zip.

Sorry for the reviewspam, we need to fix this.

Test Plan:
Unit test, built nativemsite locally.

testinprod