'buck project' builds java modules that use annotation processing
Summary:
'buck project' should set up everything so that IntelliJ has what
it needs.  When modules generate source via annotation processing,
IntelliJ needs those generated sources.  So this needs to happen
in 'buck project'.

The task for this talks about adding logic to BUILD_DEFS to do this.
However, after working through options here, it seems better to put
the fix in buck so that humans do not need to edit things in two
places (start using annotation processor in BUCK, and add target to
the initial targets list).

This change implicitly adds any java libraries that use annotation
processing to the list of initial targets and builds them in the
same pass.

Test Plan:
- existing unit tests
- 'buck project' in fb4a, verify sources are generated
- time different combinations of builds