Make it possible to disable colorization via .buckconfig.
Summary:
Now, specifying the following in .buckconfig will disable colorized output:

    [color]
      ui = false

It is helpful to be able to disable colorized output when running integration tests.

Much of this work was also modifying the code so that only one `Ansi` object is
created as part of Buck's execution and making sure it is threaded through all
of the places it needs to get to.

Test Plan:
Follow-up integration test relies on this.
Sandcastle builds.