Add a new heuristic to `buck project`.
Summary:
We recently changed `buck project` so that it only considers the first-order
deps of a build rule when determining its dependent IntelliJ modules.
This is problematic for Java packages that contain a combination of
library and test code where the set of deps of the library code is not a
subset of the deps for the test code. This diff adds a heuristic for that
case and includes the union of the deps of the library and test code in that case.

Test Plan:
Ran `buck project` and then built Messenger in IntelliJ.
Sandcastle builds.

I described a test case in the comments in the code,
but I did not create a unit test for it yet.
I am planning to finish the new integration test harness first,
and will then create a test using that. Currently, we do not
run any test code that exercises intellij.py, which is bad.