Fetch from all remotes when trying to get the required revision
Summary:
If the required revision is not in the local repository, Buck tries to
get it by running `git fetch`.  However this only works if the revision
is on the default remote, i.e. the one originally cloned from.  If the
required revision is on another remote, i.e. a forked version, it will
not be fetched.  Adding the `--all` option to `git fetch` fixes this.

Test Plan:
1. Clone buck from github.
2. Add another remote with `git remote add`.
3. In the test project set .buckversion to point to a revision that is
   only on the other remote.
4. Build the test project with `buck build`.

Change-Id: Iffd0fdff006b9bde8345d2f3623829372c1e1a83
Signed-off-by: David Pursehouse <david.pursehouse@sonymobile.com>