Have aapt ignore BUCK files in res/ directories.
Summary:
aapt will fail if it detects extra files and folders that are not
explicitly ignored. Therefore, it was previously impossible to put a
BUCK file in a res/ directory.

The motivating reason for allowing BUCK files in res/ directories is
to allow targets such as:

  # in //src/com/example/res/BUCK
  android_library(
    name = 'res',
    res = '.',
    deps = [],
    visibility = ['PUBLIC'],
  )

which could be referenced from the command line like so:
  buck build src/com/example/res

Test Plan:
Successfully ran `buck build src/com/example/res` on a project where
//src/com/example/res/ contained a BUCK file.