Create utilities to eliminate boilerplate from integration tests.
Summary:
Most integration tests need to clone a scenario folder under a relative
"testdata" directory, so TestDataHelper facilitates this.
A ProjectWorkspace can now be responsible for running Buck and reporting
the exit code, stdout, and stderr.
ProjectIntegrationTest has been simplified as a result (and no longer needs
a `tearDown` method) and serves as an example for future integration tests.

Test Plan: Sandcastle builds.