Improved heuristic for `buck project` and added a regression test.
Summary:
Libraries are supposed to be listed before modules in .iml files.
This was not necessarily the case when a test_target had a dependency
on a library rule in the same build file.

Test Plan:
Created an integration test to prevent a regression.

Note that no changes were necessary to `com.facebook.buck.command.ProjectTest`,
which implies that my changes to `Project` preserved the existing behavior for the
simpler uses of `buck project`.