Build rules bottom-up.
Summary:
Dramatically changed AbstractCachingBuildRule.build() so that a rule:

(1) Builds all of its deps.
(2) Computes its rule key.
(3) Compares its rule key to the one stored on disk.
(4) If the rule keys are different, it tries the ArtifactCache.
(5) If it has a cache miss, it builds the rule.
(6) If either (4) or (5) is true, then it updates the .success file.

Because the logic in AbstractCachingBuildRule changed so dramatically,
I commented out many of the existing tests in AbstractCachingBuildRuleTest
because they will take a considerable amount of effort to port,
but the test cases are good ones, so this needs to be done.

Still need to reinstate tests in DefaultJavaLibraryRuleTest and
AbstractCachingBuildRuleTest.

Test Plan: Sandcastle builds.