Add support for ASM in buck
Summary:
This adds the ability to specify a list of ASM commands to be run
while building a java_library() target in buck.

The command runs after all the .java files are compiled to .class
files, but before the .class files are packaged into a .jar.

The command-line is formed by splitting each command into arguments
by spaces and appending an additional argument containing the path
to the directory containing all the .class files.  The tool should
create, delete, or transform .class files in-place.

Test Plan:
- added unit test