build-time downscaling for dpi-specific builds
Summary:
Our MDPI builds contain several HDPI and XHDPI drawables. This is because no
MDPI alternatives were ever provided and, as such, the resource
filter cannot discard the larger versions.

It's not a functional problem as Android downscales automatically at runtime,
however we can save at least 0.5 MiB in APK size for MDPI builds by performing
this operation at build time.

The question is, what implementation should we provide for image scaling?
ImageMagick is free, fast and does exactly what we want, it's available in the
package managers of our build machines, but it's not installed by default.