Scaffolding for using the ABI to determine whether a rule needs to be rebuilt.
Summary:
This introduces a new interface, AbiRule, which a BuildRule can implement if
it has the potential to avoid a rebuild due to an unchanged ABI.
An AbstractCachingBuildRule will do some preliminary checks on an AbiRule to
try to avoid a rebuild.

Test Plan: Added a test case to AbstractCachingBuildRuleTest.