Change BuildRule.getOutput()->File to BuildRule.getPathToOutputFile()->String.
Summary:
Also require BuildRule.getOutputKey() to take a ProjectFilesystem.

Testing and distributed builds are easier to implement when relative paths
are used that can be resolved to absolute paths (or File objects) only when either:
(1) The path is being used by a Command that is executed outside of the project root.
(2) Information about the File is needed (whether it exists, its contents, etc.)

Test Plan: Sandcastle builds.