Integration test to verify the behavior of how timeouts affect JUnit tests.
Summary:
For a test that uses @Before and threads, specifying a timeout via @Test
is not the same as specifying a timeout via @Rule(Timeout) in JUnit according
to https://github.com/junit-team/junit/issues/686. For now, Buck faithfully
reproduces this behavior of JUnit, though we may want to change this in the
future, as this does not play well with the default timeout in .buckconfig.

Test Plan: Verified that the test passes when run from either Buck or Eclipse.