Make Verbosity a field of Console.
Summary:
Making Verbosity a field of Console will make it easier to check the
Verbosity when deciding whether to write to the Console.
Because this diff entailed adding a new required parameter to Console's
constructor, rather than add a new param to all of the instances of
"new Console" under test/, I changed them to use TestConsole instead.

Test Plan: Sandcastle builds.