Introduce AbstractBuildRuleBuilderParams for an AbstractBuildRuleBuilder constructor.
Summary:
`AbstractBuildRuleBuilderParams` is designed to serve as a context object that we can
hang things off of to pass information to all `AbstractBuildRuleBuilder`s. Previously,
`AbstractBuildRuleBuilder` used the default constructor, so we were not able to inject
any data through these objects.

Initially, `AbstractBuildRuleBuilderParams` will have only one method:

    public Function<String, String> getPathRelativizer();

By limiting this interface, it makes it easier to implement a
`FakeAbstractBuildRuleBuilderParams` that can be used in unit tests.

Most of the changes under `src/` are changes to subclasses of
`AbstractBuildRuleBuilder` and `AbstractBuildRuleFactory`.

Most of the changes under `test/` is the passing of
`new FakeAbstractBuildRuleBuilderParams` to a `Builder`'s constructor.

Test Plan: Sandcastle builds.