Allow tests included through the command line or the -i flag to be run
Summary:
If a test is excluded via "-e" or the global config, we should allow it to be run if the user
specifies it on the command line, or overrides the label with "-i". A follow up diff will add
"--only" which was the original intention of "-i".

Test Plan: buck test --all