Exclude testdata from Javadoc and fix some malformed Javadoc tags.
Summary:
In general, this diff tries to tighten up our Javadoc generation:
* Modified build.xml to fail if there are any Javadoc errors.
  Unfortunately, I have never found a nice way to treat Javadoc
  warnings as errors:
  http://stackoverflow.com/questions/8392559/javadoc-treat-warnings-as-errors
* Modified our CI scripts to run `ant javadoc`.
* Fixed the URL to link to for JUnit javadoc.
* Moved ABI test data files into a testdata directory and introduced
  the use of TestDataHelper in AbiWriterTest.

Test Plan:
Ran `ant javadoc` and verified that `com.example` no longer shows up
in the package list, and that there are no Javadoc warnings or errors.
A Javadoc error will break the CI build, but a Javadoc warning will not.