Disable artifact caching for 'buck project'
Summary:
As a temporary workaround while buck is generating annotation sources,
we need to disable artifact caching for 'buck project'.  This is needed
because the annotation processing initial targets are added to the build
purely for their side-effects.  Obtaining the cached result defeats the
purpose of adding these targets to 'buck project'.

Test Plan: - unit tests