Introduce buck.py server mode (parse perf pt2)
Summary:
Added a --server flag for buck.py which allows it to accept BUCK files to parse
on stdin so that multiple invocations of the script are not necessary to parse
a series of BUCK files.  In this mode, buck.py's output is also better structured
so that each individual BUCK file can be correlated with the list of rules
produced.

This doesn't actually improve parse time because Jython wasn't paying a
significant penalty per invocation of buck.py as its currently being used.  If
a python interpreter were to be used, however, the parse time step should
be 5x faster with this.

Test Plan: Smoke test building buck