Improve BUCK file parse error behaviour.
Summary:
Modify ProjectBuildFileParser to throw a checked exception instead of
the more generic and overloaded IOException.  Previously, all errors
from buck.py interaction were treated as an IOException and thrown
all the way to the main method.

Test Plan:
Created a syntax error in a BUCK file, observed that the output of
buck build originally was undesirable (double stack traces, the latter
of which was an I/O exception with little relevance to the actual
issue).  buck targets output is still not great (it prints a BUILD
FAILED message complaining about a missing target even though its not
the target which contained the BUCK file syntax error).  buck.py's
output is passed through to stderr and does show the parse error
plainly in either case.