Use fully-qualified names for libraries in IntelliJ configs.
Summary:
We used to do something cute where the library name would be
close to the JAR file name, which is what a human would probably do.
Unfortunately, we are starting to have situations where we have
generated JAR files with the same names in different directories,
so we were getting collisions for IntelliJ library names.

This changes things so that we do something canonical and boring,
but at least it avoids the possibility of a collision.

Test Plan:
Inspected generated IntelliJ project.
Had to update unit tests as I would expect.