`buck clean` works on Windows.
Summary:
Now `buck clean` and `buck clean --project` works on Windows. In order to do this, I have
to close the log file before cleaning, because on Windows you can not delete a file that
is opened by the current process or another process.

Test Plan: Run `buck clean` on Windows.