Create an AsyncEventBus with MoreExecutors.sameThreadExecutor() for tests.
Summary:
We have seen some flakiness in the //test/com/facebook/buck/event/listener:listener
test target because we use an AsyncEventBus to power BuckEventBus. Therefore,
we need to manually "flush" the AsyncEventBus during tests to make sure events
are pumped through so they can be verified. Our current way of doing this has
made the tests messier, required a busy/wait, etc.

Rather than use a synchronous EventBus for tests and an AsyncEventBus in prod,
this diff uses an AsyncEventBus with a same-thread Executor, which seems to do
the trick. This way, we don't need to cast an EventBus to an AsyncEventBus
anywhere in our code.

This allows us to remove a bunch of code from BuckEventBus. Getters that only
exist because they are `@VisibleForTesting` are now exposed through
`test/com/facebook/buck/event/BuckEventBusFactory.java`, which is also the
class used to create instances of BuckEventBus for testing. This ensures that
the public API of BuckEventBus remains minimal.

This diff also cleans up test/com/facebook/buck/event/BUCK, as there was a unit
test (LogEventTest) that was not being executed by Buck.

Test Plan:
Ran `buck test test/com/facebook/buck/event/listener` 10x from the command line.
Did not observe any failures.