Delete stale metadata when an artifact is loaded from the cache.
Summary:
This diff fixes `DefaultJavaLibraryRule` so that it deletes metadata files
when an artifact is fetched from the cache to eliminate potential
corruption issues. Unfortunately, the absence of this metadata means
that future builds of the `DefaultJavaLibraryRule` rule may be suboptimal,
as the rebuild could be avoided if the metadata for the artifact were known.
Currently, we need a short-term fix, but the correct long-term fix is to pull
down the new metadata with the artifact so that the metadata on disk
can be replaced instead of categorically deleted.

Ultimately, we are going to create a framework for storing and loading
additional files and data for a build rule so that such transactions with
the artifact cache are atomic. Hopefully this will eliminate an entire class
of corruption issues.

Test Plan:
Wrote `CachedTestIntegrationTest` first and verified that it failed.
However, when `mode = dir` was commented out in the `.buckconfig`
file, the test would pass. This demonstrated that the problem was
due to a bad interaction with the build cache.

Then made the fix in `DefaultJavaLibraryRule` and verified that the
test passed, as expected.