Move getPathToOutputFile() from BuildRule to Buildable.
Summary:
This introduces 6 casts to `Buildable`.
It's dirty, but I believe it is a necessary interim step.

Ultimately, we are trying to move to a world where a rule
declares all of the files it will produce, all of which are
tar'ed up and stored in the artifact cache as a unit.

Test Plan: Sandcastle builds.