Ignore sh_test on Windows.
Summary:
Shell script is not executable on Windows.
Ignore this kind of test on Windows by always returning success.
We would like to add another parameter to `sh_test` (for example `test_win`)
or another function (for example `cmd_test`) to support
the same thing on Windows in future.

Test Plan: Unit tests.