Make Buildable a field of the AbstractCachingBuildRule.
Summary:
Next step on the path to making AbstractCachingBuildRule "a thing that runs
Buildables". Make the Buildable a field on the class, but try and keep as much
of the rest of the code base as stable as possible. We do this by introducing
two new classes.

The first is DoNotUseAbstractBuildable, which is a simple extension of
AbstractCachingBuildRule, and provides default implementations for most of the
Buildable interface, allowing us to remove those from ACBR.

The second is the same, but without using inheritance: it's just an implementaion
of Buildable. Its use is demonstrated by migrating ExportFile to a plain
Buildable.

Test Plan:
buck test --all
Ensure that our prod apps compile and run on an emulator