Hack ProjectWorkspace on Windows.
Summary:
Some unit tests rely on `junit.jar` which is a symbolic link in the respository.
Symlinks are not supported directly on Windows, so that they are checked out
as normal files containing a one-line path, which are hard to diffrentiate from real normal files.

Here is a hack: Check if copied file is "like" a symbolic link. If the file is small and contains
a one-line path, then we think it is a symbolic link and copy the target.

Test Plan: Unit tests.