Add platform-specific command parameters to genrule.
Summary:
Add 2 parameters `win` and `sh` to `genrule` and make `cmd` optional,
in which `win` is Windows-specific and `sh` is non-windows specific command.
`cmd` is platform independent and has a lower priority if the platform-specific command is given.

The priority sequence is (on Windows):

  "cmd.exe /c [cmd_exe]"
  "cmd.exe /c [cmd]"

On non-windows os:

  "/bin/bash -c [bash]"
  "/bin/bash -c [cmd]"

This would provide backward compatibility.

Test Plan: Unit tests.