Fix the computation of the hash for a dirty Git repository.
Summary:
Previously, the call to `git ls-tree` to compute the hash of a dirty
Git repository specified a list of paths to include in the hash.
This list was:

  bin build.xml config lib third-party

The `src` directory and `DEFS` and `.buckconfig` files were conspicuously
missing from this list, as they all affect the behavior of Buck,
as the files in the list do.

However, it is simpler and safer to consider all version controlled files
when computing the hash, which is what `git ls-tree` without any paths
appears to do.

Test Plan:
I took the appropriate chunk of `bin/buck_common` and ran `git ls-tree`
without any paths. The contents of the resulting file were:

```
100644 blob 269c00a240c000ba34d1ed65c121f497cedcab6d	.arcconfig
100644 blob d22dc6e7f9d353c6e8a2c128dc35cfbe36e42ae3	.buckconfig
100644 blob 192a018b263c876bee0bc6ed5412fde721ce5f0c	.classpath
100644 blob 463747e3d3f7fe70afba1711a11f9453e3cbb113	.gitignore
040000 tree 960bc780d7fdbe18d2f5871cc9d23f73a16ee9e6	.idea
100644 blob b86ac4ac38bff4e1fe13eb6f1bd23108d7a2da50	.project
100644 blob 742e765d4d21710a7f517aee061dc98f00215f6b	DEFS
100644 blob d645695673349e3947e8e5ae42332d0ac3164cd7	LICENSE
100644 blob 42ed037e0c84bdf2cc65b2802e75af198f4de225	NOTES.txt
100644 blob 962f05867c923e6c0caa0872c3fba84f6178efc9	README.md
100755 blob e88a1affe24d4b4ce7179287c32150bbb091d1db	arc_unit.sh
040000 tree c42fc50ae9028f5ddf52759a0489b7f7c377f25f	arcanist
100644 blob a4a7a21d544a57dece2d315129565438a4f1c309	assert_code_coverage.py
100755 blob 4734d04c83703843929230d3dd6fb1de010b6995	automation.sh
040000 tree cca72d3917b3abe6f4d9dc338eceeee14d3b218c	bin
100644 blob d1b85a4348dce1a4cb2ecd5ad59ccad88106579c	buck.iml
100644 blob 81f7f4fcd81884a5aefa12b2e432b125021a64a4	build.xml
040000 tree 5d333f56235a5d20f6cdccff205b6556ef8fc8ca	config
040000 tree a1bcb00b5e48b4d230e8488d73349221d9ff8235	docs
040000 tree a5f9be5be8daf0dc542149dad8465c96a078f4e2	lib
040000 tree 8c86bb01a6f8515f76c9cdfb22516ab029fd2ae7	opensource
040000 tree 963ba362422381b820146f6a5b44bf660c4ee0c0	pmd
040000 tree 16d2973d11bd69abfe0c6621b5ae865c6037da20	scripts
040000 tree fe7d5fb59b075d0aa4f8867a4dcd5a0ea6fb804c	src
040000 tree 96c9fa2f6b0a1f4d5f684b5c54f8a635ee9426de	test
040000 tree 07d98f0dbb37b0c15346affe5f3f25f1cfa616f7	testdata
040000 tree 93be27eed3fb2667aa2116830615151dd8d74e1b	testrunner
040000 tree 964cd008c6a9daf6f7072b5b47cc69823b32d998	third-party
```

Note that `buck-out` and `build` are omitted from this list even
though they exist in the local Buck repository. This demonstrates that
generated files are excluded from the repository hash, as desired.