Add use_linear_alloc_split_dex option to android_binary().
Summary:
If `use_linear_alloc_split_dex` is passed to `android_binary()` and
`use_split_dex` is `True`, then use `LinearAllocEstimator` to determine
how secondary dexes should be spilt.

This diff changes `ZipSplitter` into an interface, so the old `ZipSplitter` code is
divided between `DefaultZipSplitter` and `AbstractZipSplitter`.

As the point of this diff is to introduce new zip-splitting logic,
it introduces a new implementation of `ZipSplitter` named
`LinearAllocAwareZipSplitter` that uses `LinearAllocEstimator` to determine
how to divide class files into groups that can be zipped.

This also updates a `ZipSplitter` so that it generates a report of which files
were written to which zips and what their size estimates were.