Implemented a Websocket server that reports events.
Summary:
In order to get the progresses of build from external process,
an interface is needed to collect all internal events and export
them interactively. We implemented a Websocket server based on jetty,
which provides a universal API. We can retrieve the progresses from both Java and browsers.

Test Plan:
Run `buckd` and create a Websocket client from browser and connects to
http://localhost:7013/comet/echo. Then run `buck test --all` to see the events.