Clean up temp files if there is no cache result.
Summary:
We create a temp file to uncache files to. However, if we miss the cache entirely, there
is no hit in the cache at all, we leave a zero-length file in the temp directory. Let's
clean that up to avoid cluttering the disk.

Test Plan:
buck test --all
Manually inspect the tmp dir for zero-length zip files