Add a BUCK_COLOR env var, fixing SuperConsole in buckd.
Summary:
buckd can't tell whether the caller is talking to a tty or not, so we need to
tell it in order to enable SuperConsole output. Add BUCK_COLOR, which
(extending the existing color config) takes values of true/always, false/never,
or auto. Copy the VerbosityParser hack to get this option early in startup.
Modify the buck starter script to pass this value to the nailgun server.

Thoughts on renaming the "color" config to "fancy" or something?

Test Plan: buckd output is now Super by default. Confirm that "never" overrides the default.