Fix AndroidBinaryRule.Builder so that build() can be invoked multiple times.
Summary:
A previous diff introduced a change to `AndroidBinaryRule.Builder` such that invoking
`build()` would leave around some state (only if pre-dexing was applied) that
would cause a problem if `build()` were invoked again on the same `Builder` object.

When `buckd` is not running, `build()` is only invoked once, so this went
undetected in the original diff.