Declare daemon as volatile
Summary:
The getDaemon() method contains an unsynchronized lazy initialization
of a non-volatile static field. Because the compiler or processor may
reorder instructions, threads are not guaranteed to see a completely
initialized object, if the method can be called by multiple threads.

Make the field volatile to correct this potential problem.

Test plan:
Run test builds.

Change-Id: I6f47da94f76d988649f411121772f372743f3b77