Rename the ArgObjectPopulatomatic to ConstructorArgMarshaller.
Summary:
When this class was added, it wasn't clear what a sensible name for it would be, and so
a totally unsuitable and ludicrous one was picked. As work on the plugin API has
progressed, the intent of this class has become clearer, and a better name can now be
selected.

Test Plan: buck test --all