Support multiple target densities.
Summary:
We used to focus on MDPI, HDPI and XHDPI, and we only
had drawables of that density. As we're now including XXHDPI drawables,
it's time we filtered them from MDPI and HDPI builds.

This adds support for specifying multiple target densities explicitly.
Ideally, with both diffs in, the only change is, we won't have LDPI drawables in HDPI, TVDPI and XHDPI, XXHDPI builds,
and we won't have XXHDPI drawables in anything but the XHDPI, XXHDPI build.