Remove custom ExtractResourcesStep in lieu of ApkBuilder.addResourcesFromJar().
Summary:
Historically, if we needed to include a java_library that had resources
(i.e., files that it loaded via `Class.getResource()`), then we would have
to package those files into the APK so that `Class.getResource()` worked
when that code was run in an APK just like it did in a JAR.

To achieve this, we previously used `ExtractResourcesStep` to extract all
of the resources from all of the JAR files that went into an APK, and wrote
the resources to a directory. Then we passed that directory to `aapt`.

Back when I was still figuring out how the Ant scripts worked, I noticed that Ant
did this automagically. Presumably it was using this feature of `ApkBuilder`, which
I duplicated with my `ExtractResourcesStep`. Letting `ApkBuilder` do this seems
like a better way to go.

Most importantly, for my current purposes, this allows us to defer the computation of
`dexTransitiveDependencies` in `AndroidBinaryRule#getBuildSteps()`, which will make
it easier to deconstruct that method and farm it out into more specific build rules.