`InitializableFromDisk` must declare the type of object it populates.
Summary:
Historically, the burden has been on developers to remember to implement
`initializeFromDisk()` in order to put in-memory data structures in the same
state when reading a `BuildRule` from cache as it would have created them when
building the rule locally. This was problematic as it required developers to
implement two code paths to achieve the same result, which was error-prone.

Here, we eliminate this problem by forcing this logic to go through one code path.
This diff requires the steps to serialize any data structures to disk,
and then AbstractCachingBuildRule will invoke `initializeFromDisk()`,
which must now return the data structures created anew, having been read from disk.
This will help with correctness, though it may also make things slightly slower,
as data that was previously immediately available to a `BuildRule` in memory now
has to be serialized and deserialized before it can be used.

Test Plan: Sandcastle builds.