Fix SuperConsole thread leak.
Summary:
Previously each buck command potentially created a
SuperConsole instance which created a ScheduledExecutorService which
was then leaked when running the command inside buckd. This change
makes the ScheduledExecutorService static, the SuperConsole Closable
so that it can cancel its scheduled updates and creates the
SuperConsole inside a try-with-resources block around command execution so
that it is always cleaned up properly.

Test Plan:
0) buck test --all
1) buckd
2) for DIR in `find . -type d`; do buck query "$DIR ->"; done
3) killall -3 java
4) grep SuperConsoleEventBusListener .buckd/buckd.log
5) check that only a single SuperConsoleEventBusListener thread is shown in 4)