Do some performance tuning for methods that were taking a lot of CPU according to jvisualvm.
Summary:
This reduces the overhead of the `equals()` and `hashCode()` methods of `AbstractBuildRule`.
Apparently we spend a lot of time putting `BuildRule` objects into collections, so making
these methods more efficient should help.

This diff also removes a precondition that appears to have a non-trivial cost.
In practice, `BuildRule.getDeps()` will often return an object of type
`com.google.common.collect.RegularImmutableSortedSet`. Its `contains()` method
was also showing up as taking up a lot of CPU time in jvisualvm. When I looked
at the implementation, it invokes a method named `unsafeBinarySearch()`,
which does not sound good to me. We should probably investigate this further,
but for now, I am going to remove this precondition.

Test Plan: Sandcastle builds.