Move the logic for generating the final R.java for the APK into a separate build rule.
Summary:
This diff requires a considerable amount of refactoring to `AndroidBinaryRule` to
untangle the dependencies so that R.java can be generated and compiled in a
separate rule. Part of this complexity stems from the resource filtering that we
do for both drawables and strings, as it conditionally affects the set of res/
directories to pass to `aapt`.

Test Plan: Sandcastle builds.