ignore -nodpi drawables from filtering
Summary:
Android doesn't try to scale these, we probably shouldn't either.

Simplest workaround for not being able to define a `messenger_ldpi` target,
where the previous version of the code tried to scale to a `drawable-nodpi-ldpi`,
because it treated `drawable-nodpi` as a baseline 160dpi because it had no
other density qualifiers.