Fold AccumulateClassNames into JavaLibraryRule.
Summary:
In practice, `AccumulateClassNamesStep` is very fast, so having a
separate `Buildable` for `AccumulateClassNames` creates unnecessary overhead
because it results in more Cassandra lookups.

It was also causing unnecessary rebuilds. For example, a `JavaLibraryRule` might
avoid rebuilding itself due to `MATCHING_DEPS_ABI_AND_RULE_KEY_NO_DEPS`.
However, the `AccumulateClassNames` that depends on it would still be rebuilt
locally because its `RuleKey` had changed (most likely due to some ancestor
`JavaLibraryRule` that had to be recompiled).

Normally, we would avoid this sort of rebuild by making `AccumulateClassNames` an
`AbiRule`; however, its dep (the `JavaLibraryRule`) already has an ABI key, which
is not the same key that `AccumulateClassNames` would need to determine whether it
needed to build itself. Morever, the ABI key that it would need from the
`JavaLibraryRule` would be the one it was designed to compute!

Overall, the simplest solution was to accumulate the `.class` file names and hashes
as part of `JavaLibraryRule`. This should speed up builds by avoiding unnecessary
rebuilds.

A side-effect of this diff was consolidating some logic common to
`DefaultJavaLibraryRule` and `PrebuiltJarRule` in a new `JavaLibraryRules`
utility class.

Test Plan: Sandcastle builds.